[SQL/Translator/{Parser,Producer}/PostgreSQL] support index expressions
It would be nice the return the expression as scalar references, but
SQL::Translator::Schema::Index::new stringifies them.  It also splits
on commas, so we need instead avoid quoting expression parts that have
"expression characters" like parens and space and hope that the
producer will put the expressions back together correctly. So

   CREATE INDEX ON play (a, lower(b, c, d), e desc, (this + that))

will result in a fields list of

   ['a','lower(b', 'c', 'd)', 'e desc', '(this + that)']

which when join ", " is applied will give back the correct expression.
It's odd, I must say, but it works.  Notice that the code API can
use

    fields => ['a', 'lower(b, c, d)', 'e desc', '(this + that)']

and it will be auto-split into the above.  Notice that we have to
avoid quoting expressions that match /[( )]/ for this to work.